refactor: move types and interfaces files to types folder
