+++ src/app.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { User } from './db/types';
+import { User } from './types/types';
 
 @Injectable()
 export class AppService {
